requirements.txt
requests==2.32.3
pyarrow==17.0.0
google-cloud-storage==2.18.2
google-cloud-secret-manager==2.20.2
